kafka/requirements.txt
kafka-python-ng>=2.2.0
pandas==2.1.4
python-dotenv==1.0.0
redis>=5.0.0
pytest>=7.4.0

